refactor(graphql-client): reorganize faucet and output types
Refs #512.
