+++ crates/iota-sdk-graphql-client/src/output_types/transactions.rs
@@ -1,0 +1,13 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use iota_types::{SignedTransaction, TransactionEffects};
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+pub struct TransactionDataEffects {
+    pub tx: SignedTransaction,
+    pub effects: TransactionEffects,
+}
+
